refactor(oceans-e2e): move POMs into e2e/poms/ subdirectory
Keeps spec files at e2e/ root and page objects in e2e/poms/ so the
two concerns are visually separated. All 41 tests pass.
