+++ frontend/packages/labs/oceans/e2e/poms/FishVTrashPage.ts
